+++ keras_hub/src/models/mistral/mistral_transformer_decoder.py
@@ -1,6 +1,5 @@
 import keras
 from keras import ops
-from keras.layers import RMSNormalization
 
 from keras_hub.src.layers.modeling.transformer_layer_utils import (
     compute_causal_mask,
@@ -10,6 +9,9 @@ from keras_hub.src.layers.modeling.transformer_layer_utils import (
 )
 from keras_hub.src.models.mistral.mistral_attention import (
     CachedMistralAttention,
+)
+from keras_hub.src.models.mistral.mistral_layer_norm import (
+    MistralLayerNormalization,
 )
 from keras_hub.src.utils.keras_utils import clone_initializer
 
@@ -69,7 +71,7 @@ class MistralTransformerDecoder(keras.layers.Layer):
         )
         self._self_attention_layer.build(decoder_sequence_shape)
 
-        self._self_attention_layernorm = RMSNormalization(
+        self._self_attention_layernorm = MistralLayerNormalization(
             epsilon=self.layer_norm_epsilon,
             dtype=self.dtype_policy,
             name="self_attention_layernorm",
@@ -114,7 +116,7 @@ class MistralTransformerDecoder(keras.layers.Layer):
             )
         )
 
-        self._feedforward_layernorm = RMSNormalization(
+        self._feedforward_layernorm = MistralLayerNormalization(
             epsilon=self.layer_norm_epsilon,
             dtype=self.dtype_policy,
             name="feedforward_layernorm",
