+++ keras_hub/src/models/mistral3/mistral3_vision_encoder.py
