+++ keras_hub/src/utils/transformers/preset_loader.py
@@ -24,6 +24,7 @@ from keras_hub.src.utils.transformers import convert_gpt_oss
 from keras_hub.src.utils.transformers import convert_llama3
 from keras_hub.src.utils.transformers import convert_metaclip_2
 from keras_hub.src.utils.transformers import convert_mistral
+from keras_hub.src.utils.transformers import convert_mistral3
 from keras_hub.src.utils.transformers import convert_mixtral
 from keras_hub.src.utils.transformers import convert_pali_gemma
 from keras_hub.src.utils.transformers import convert_qwen
@@ -83,8 +84,10 @@ class TransformersPresetLoader(PresetLoader):
             self.converter = convert_llama3
         elif model_type == "metaclip_2":
             self.converter = convert_metaclip_2
-        elif model_type in ("mistral", "mistral3"):
+        elif model_type == "mistral":
             self.converter = convert_mistral
+        elif model_type == "mistral3":
+            self.converter = convert_mistral3
         elif model_type == "paligemma":
             self.converter = convert_pali_gemma
         elif model_type == "vit":
